--- apps/examples/src/examples/custom-shape/snapshot.json
@@ -1,92 +1,0 @@
-{
-	"store": {
-		"document:document": {
-			"gridSize": 10,
-			"name": "",
-			"meta": {},
-			"id": "document:document",
-			"typeName": "document"
-		},
-		"page:page": {
-			"meta": {},
-			"id": "page:page",
-			"name": "Page 1",
-			"index": "a1",
-			"typeName": "page"
-		},
-		"shape:Q3Q3RJnUTlggH_00r38Hu": {
-			"x": 100,
-			"y": 100,
-			"rotation": 0,
-			"isLocked": false,
-			"opacity": 1,
-			"meta": {},
-			"id": "shape:Q3Q3RJnUTlggH_00r38Hu",
-			"type": "my-custom-shape",
-			"parentId": "page:page",
-			"index": "a1",
-			"props": {
-				"w": 200,
-				"h": 200,
-				"text": "I'm a shape!"
-			},
-			"typeName": "shape"
-		}
-	},
-	"schema": {
-		"schemaVersion": 1,
-		"storeVersion": 4,
-		"recordVersions": {
-			"asset": {
-				"version": 1,
-				"subTypeKey": "type",
-				"subTypeVersions": {
-					"image": 3,
-					"video": 3,
-					"bookmark": 1
-				}
-			},
-			"camera": {
-				"version": 1
-			},
-			"document": {
-				"version": 2
-			},
-			"instance": {
-				"version": 24
-			},
-			"instance_page_state": {
-				"version": 5
-			},
-			"page": {
-				"version": 1
-			},
-			"shape": {
-				"version": 3,
-				"subTypeKey": "type",
-				"subTypeVersions": {
-					"group": 0,
-					"text": 1,
-					"bookmark": 2,
-					"draw": 1,
-					"geo": 8,
-					"note": 5,
-					"line": 4,
-					"frame": 0,
-					"arrow": 3,
-					"highlight": 0,
-					"embed": 4,
-					"image": 3,
-					"video": 2,
-					"my-custom-shape": 0
-				}
-			},
-			"instance_presence": {
-				"version": 5
-			},
-			"pointer": {
-				"version": 1
-			}
-		}
-	}
-}
