+++ apps/examples/src/examples/editable-shape/EditableShapeExample.tsx
@@ -1,18 +1,16 @@
 import { Tldraw } from 'tldraw'
 import 'tldraw/tldraw.css'
-import { MyShapeUtil } from './MyShapeUtil'
+import { EditableShapeUtil } from './EditableShapeUtil'
 
-// [1]
-const customShapeUtils = [MyShapeUtil]
+const customShapeUtils = [EditableShapeUtil]
 
-//[2]
 export default function EditableShapeExample() {
 	return (
 		<div className="tldraw__editor">
 			<Tldraw
 				// Pass in the array of custom shape classes
 				shapeUtils={customShapeUtils}
-				// snapshot={snapshot}
+				// Create a shape when the editor mounts
 				onMount={(editor) => {
 					editor.createShape({ type: 'my-editable-shape', x: 100, y: 100 })
 				}}
@@ -26,24 +24,12 @@ Introduction:
 
 In Tldraw shapes can exist in an editing state. When shapes are in the editing state
 they are focused and can't be dragged, resized or rotated. Shapes enter this state 
-when they are double-clicked, this means that users can drag and resize shapes without 
-accidentally entering the editing state. In our default shapes we mostly use this for 
-editing text, but it's also used in our video shape. In this example we'll create a 
-shape that you could use for a game of Go, but instead of black and white stones, we'll
-use cats and dogs.
+when they are double-clicked. In our default shapes we mostly use this for editing text. 
+In this example we'll create a shape that renders an emoji and allows the user to change 
+the emoji when the shape is in the editing state.
 
-Most of the relevant code for this is in the my-shape-util.tsx file. We also define a
-very simple tool in my-shape-tool.tsx, and make our new tool appear on the toolbar in 
-ui-overrides.ts. 
+Most of the relevant code for this is in the EditableShapeUtil.tsx file. If you want a more
+in-depth explanation of the shape util, check out the custom shape example.
 
-[1]
-We have to define our array of custom shapes and tools outside of the component. So it
-doesn't get redefined every time the component re-renders. We'll pass that in to the 
-editors props.
-
-[2]
-We pass in our custom shape classes to the Tldraw component as props. We also pass in
-any uiOverrides we want to use, this is to make sure that our custom tool appears on 
-the toolbar. 
 
  */
